AJCORE-1657 Add private copy/assign to ECCSecret class
When a class or struct has dynamic memory it should also
define a copy constructor and assignment operator to prevent
double freeing of memory.

If the clas or struct should never be copied a private version
of these functions must be added to prevent the compiler from auto
generatign the functions.

Change-Id: I16f7dabe2ea4139e13649d23d20515d12ba4510d
Signed-off-by: George Nash <georgen@qce.qualcomm.com>