ASACORE-1736 fix ObserverManager segfault
ObserverManager was too enthousiastic in its shutdown sequence: work
items should not be deleted from the queue before the last in-flight
work item has landed. A side effect of purging the queue is deleting any
unregistered CoreObservers, which may go wrong if the in-flight work
item references one of these.

Change-Id: Ie96c558467ca243f35df1ad88901217985cacded
Signed-off-by: Meggan Ehret <meggan@qeo.org>