AJCORE-1666 Add private copy/assign to Crypto_ECC class
When a class or struct has dynamic memory it should also
define a copy constructor and assignment operator to prevent
double freeing of memory.

If the clas or struct should never be copied a private version
of these functions must be added to prevent the compiler from auto
generatign the functions.

Change-Id: Icef61837d98cf53c2fe84ccb6178f853fb045d8b
Signed-off-by: George Nash <georgen@qce.qualcomm.com>