ASACORE-1734 Timer thread on Windows may deadlock
There are a couple of places where the decision to add a new
alarm depends on thread's name, which is platform specific.
This change removes that dependency and prevents the thread
currently servicing the alarm from adding another one and
effectively waiting on itself.

I also fixed these issues uncovered in stress tests:
- Trim thread pool memory right away after alarm completes.
- It is incorrect to wait in Stop(), so move that into Join().
- An alarm can get removed from the timer by its own handler
using the blocking version of the API. In this case, make an
exception that thread should not be blocked.
- Only block Remove and Replace operations if and only if the
matching alarm is already in flight.

Note:
- There are still a couple other known deadlocks. See
    - ASACORE-1755 (deadlock in name service).
    - ASACORE-1786 (deadlock in nullEndpoint).
- While working this bug, found potential improvement in posix
timer. See ASACORE-1762.

Change-Id: I02eee5e94887e5d129b347803e34da97d121a149
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>