ASACORE-1817 Fixed issue that was causing test programs to not Announce
The test programs and samples has a local instant of the AboutObj.
When the object was cleaned up the announce signal is also canceled.
this caused samples that were waiting for that sample to not work.

bastress2 was missing the default language in it AboutData and this
was causing the announce method call to fail.

Change-Id: I4de60c600fab3f214a23b00082eae56e81be0d03
Signed-off-by: George Nash <georgen@qce.qualcomm.com>