ASACORE-1820 Compile error for AllJoynChat sample when built for 64-bit devices
The chat sample had a debug log which prints out the value of a pointer. While
this worked great on a 32-bit device it created a possible loss of information
when compiled for a 64-bit device. The pointer type was changed from uint to
uintptr_t as per the guidelines of migrating an app for 64-bit device

Change-Id: Ibe4f3dd9d33257656fa9620ef93d27c8ade14f37
Signed-off-by: Nikhil Dabhade <ndabhade@qce.qualcomm.com>