Fix the test code crash when compile with CRYPTO=builtin
The DRBG context object was null when it is accessed by the
Random generator code.

The test code is changed to all AllJoynInit().
Change the usage of GUID128 as static variable with default constructor
because the GUID128 default constructor is calling the DRBG code
before it is initialized by AllJoynInit().

Change-Id: Id6b51a8adeb5688a32a2c2728d4cb2d937c7d252
Signed-off-by: Phil Nguyen <philn@qce.qualcomm.com>