ASACORE-1881 Revert Windows Timer for Posix
The current Windows timer does not guarantee firing in order (if alarms are
of the same expiration time). The implementation relies heavily on the OS
thread pool, which does not behave that way.

Because it is fundamental to AllJoyn, introducing such synchronization is
too risky at this point as proven by the length of time ASACORE-1758 took.

This change is to simply revert the Windows Timer back to the POSIX's
implementation.

Change-Id: Ie103006447e51c5584779629d4a732cfcbda5ef2
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>