ASACORE-1889: AJCTest SignalsTest.register_unregister_sessionlesssignal fails
ASACORE-1889: AJCTest SignalsTest.register_unregister_sessionlesssignal fails with daemon router

This test registers a sessionless signal and then immediately tries to delete
the signal. There is no wait between the announcement of the sessionless signal
and the cancel call, so while it passes with a bundled router this test often
fails with the daemon router on Windows 10. Adding a small sleep allows the
test to consistently pass.

I also opened ASACORE-1894 to track the underlying product issue with this
race condition.

Change-Id: I6dba8aa40125cfdd7e1b1681cae7b3ec691f3b94
Signed-off-by: Rob Smith (WINDOWS) <robs@microsoft.com>