ASACORE-1890: C runtime on Windows uses 32 bit time for x86 apps
The in-built Windows C runtime uses 32 bit time for x86 apps, not 64 bit time. This
is not the default behavior in applications built with Visual Studio, but appears
to be the case for OS applications. Specifically the _ftime and gmtime functions are
being mapped to the 32 bit versions for x86 builds of AllJoyn.

The use of 32 bit time causes dates after the year 2038 to fail, including the
following test case: CertificateECCTest.X509CertExpiresBeyond2050. This is a product
bug that would cause cert validation issues for any certificate which was valid
beyond the year 2038.

AllJoyn already contains a time class in the PAL, the time class needs to be updated
to explicitly call the 64 bit versions of these methods. There are also instances in
the AllJoyn code that are calling the time methods directly instead of using the PAL,
and those need to be re-routed to call through the PAL.

Change-Id: I597d830180fa8e4337a24a054676839f30532fd4
Signed-off-by: Rob Smith (WINDOWS) <robs@microsoft.com>