Use a composite key in the key store
Replace the GUID with a composite key in the key store.
The composite key is the GUID plus a type to allow for locally
and remotely generated GUIDs to be stored in the key store
without overwriting each other.

This would prevent attack by peer to overwrite key entries generated
locally.

Change-Id: If0e298c56e4ef097372498d0660442b174b1a4a6
Signed-off-by: Phil Nguyen <philn@qce.qualcomm.com>