ASACORE-1905 Use the baCommonLock in place of the jaboutObjGlobalRefLock
This is done to prevent a race condition bug that the aboutObj is checked
to see if it is null in the JBusAttachment.Disconnect method before
releasing the aboutObjectGlobalRef.  The is a small window that the
aboutObj could be deleted by the Finalize method when trying to release the
jaboutObjGlobalRefLock causing a segfault.

The lock was also added around the code that sets the aboutObj in the
JBusAttachment to null in the AboutObj_destroy function. This is done
to prevent the the destruction of the AboutObj while the BusAttachment
is in the middle of a disconnect call.

Change-Id: I23ab975a7c7f4b995a2af5af4ea3107e9bbad7ec
Signed-off-by: George Nash <georgen@qce.qualcomm.com>