Unit tests that were writen to test the legacy AboutService
These unit tests were writen quite a while ago but were not
commited because they were originally planed to be expanded
to cover more of the AboutService features.

Since these were writen we have depricated the AboutService
I have no plans to continue adding more tests but felt it
was wise to checkin the unit test for testing since the code
already exists and does add to the test verification.

Change-Id: I839bd586ddd0b94670bc4edb83b56fa0802b0935
Signed-off-by: George Nash <georgen@qce.qualcomm.com>