ASACORE-1978: AllJoyn returns incorrect error if named pipe transport fails
If a developer doesn't include the allJoyn capability in their app manifest
then connecting to the named pipe will fail with ERROR_ACCESS_DENIED.
However, this error ends up being overwritten by the status returned from
the bundled router transport, so the developer has no idea why the
connection failed.

We need to change the transport to return a better error than it currently
does, and modify the BusAttachment code to return the error for the
requested connection if connections are unsuccessful, instead of the error
from the bundled router attempt.

Change-Id: Ib9a08215b8367f85705312d23fbce73b82c35dc6
Signed-off-by: Rob Smith <robs@microsoft.com>