ASACORE-1984 Add the -Wunused-parameter flag to Linux build
The Windows build fails when there is an unused parameter, however
the Linux build does not fail when there is an unused parameter.
This makes it difficult to know the build will fail when working
on Linux. Adding the -Wunused-parameter flag to the Linux build will
cause the Linux to fail when there is an unused parameter similar to
windows. Making the build warnings similar across the OSs will help
reduce developer issues when committing code for the verification
build.

Due to C98 compilers the QCC_UNUSED is placed at the end of variable
declarations in C functions not at the top of the functions like done
elsewhere.

Change-Id: I04fd612a995e6262b91c2f40b4931b15991e1fee
Signed-off-by: George Nash <georgen@qce.qualcomm.com>