ASACORE-1988 Use uint8_t array for CertificateX509 serial number
Currently CertificateX509 uses qcc::String for the serial number.
This is misleading, because it's defined in the standard as an
integer of arbitrary size, and should not be thought of as a string
which implies human-readable. In fact, best practice for serial
numbers is for them to be random, which will result in control
characters and other unprintable characters if interpreted as a
string.

This change instead stores and treats the serial number as a uint8_t
array of arbitrary size with needed changes to sample and test code.
Some of the test code uses strings to generate bytes for a serial
number, but now they explicitly translate those strings into uint8_t
arrays and back as needed.

Change-Id: Ib9d2e84e4cfcb4cbfb2ebb917066ad12606cb55e
Signed-off-by: Kevin Kane <kkane@microsoft.com>