ASACORE-1989 Introduce cross-platform constant for size_t printfs
Some test code uses the "%zu" printf specifier, which is intended
for printing size_t (which varies depending on the platform).
%zu was introduced in C++99 and Microsoft's compilers/CRT don't
support it; instead there's an MS-specific extension called "%Iu".

There isn't a standard <inttypes.h> constant for it either, so this
change introduces an inttypes-style PRIuSIZET (for size_t) and
PRIiSIZET (for ssize_t), and uses them when printf'ing size_t's.

Change-Id: I2f876ffb717814c29554f573cfe43acf2342d429
Signed-off-by: Kevin Kane <kkane@microsoft.com>