ASACORE-2003 Encapsulate ECCPublicKey/ECCPrivateKey as classes
Change ECCPublicKey and ECCPrivateKey to be classes instead of
structs and change all code to access them to use methods instead
of direct references to internal data structures.

Change functions and methods that deal in uint8_t arrays for key
material to instead use these class abstractions.

Remove all fragile, unsafe casts of these types to the types of
their member fields so code using these types is not dependent on
the internal implementation, but only on the exported API.

Remove dependencies on ECC_COORDINATE_SZ constant as much as
possible in favor of APIs that provide sizes to allow for other
curves in the future.

Add a destructor to ECCPrivateKey to secure clear the d array.

Change-Id: Ib234e719ff49eed0e2c66c622cb8723abaec8025
Signed-off-by: Kevin Kane <kkane@microsoft.com>