Removed comparisons of unsigned values against `0`
If comparing an unsigned value against 0 the comparisons have no
effect.

From StringUtil.cc:
The method CharToU8 returns uint8_t so 0 > CharToU8(...) is always false.
In the U8ToChar variable d is uint8_t so d >= 0 is always true.

From utility.cc:
charLen is size_t so it will charLen < 0 comparison is always false the
code in that if block is never run.
mbstowcs function returns size_t so mbstowcs(...) < 0 is always false the
code in that block is never run.

This commit is removing dead code that is never run or unneeded comparisons.

Issues found using static analysis tool.
 
Change-Id: If37b4ad632f560e446d2424ed6d39e071a8b66ac
Signed-off-by: George Nash <georgen@qce.qualcomm.com>