ASACORE-2048 Remove deprecated SessionLost callback
The older version of SessionLost has been documented as deprecated for a long
time and has been superseded by a newer SessionLost version having an additional
parameter. A natural place for an app to destroy its SessionListener object is
in the SessionLost callback. However, the current BusAttachment implementation
blocks that kind of clean-up, in its effort to support both SessionLost
versions. Therefore the deprecated version of SessionLost is being removed here,
thus unblocking proper SessionListener clean-up in applications.

Signed-off-by: Daniel Mihai <Daniel.Mihai@microsoft.com>
Change-Id: I1ce92892c5f15611cbd3d97d454f5c7c1585a842