ASACORE-1578 Reworked the PermissionPolicy so it does not free user memory
The Permission Class was reworked so when a pointer to an array of peers, rules, or
members the class will no longer take over the users memory.  This is important
it helps prevent situations where memory is freed twice or memory leaks are caused
simply because the user does not know they are responsible for freeing the memory.

Notable changes.  The Peer class member KeyInfoECC was changed to KeyInfoNISTP256.
This was done since the value is used as a KeyInfoNISTP256 ins multiple places.
If it is not a KeyInfoNISTP256 then information is lost since only the KeyInfoECC
was copied into the class.

Moved most implementation details into the .cc file. Got tired of small changes
causing large recompilations since they were in the .h file.

Change-Id: I67e97cc4cbf04e2a64e0aa6950541d6ad36f0992
Signed-off-by: George Nash <georgen@qce.qualcomm.com>