ASACORE-1967 ASACORE-1702 Make property cache work with global broadcast PropertiesChanged signals
This change makes the property caching feature work well with
PropertiesChanged signals that are emitted via global broadcast instead
of sessioncast.

Because we want to avoid making PBO::EnablePropertyCaching and
PBO::AddInterface blocking calls, the setup for property caching (i.e.
adding the appropriate match rules to the routing node) is done
asynchronously. The trade-off here is that the property cache is not yet
enabled when PBO::EnablePropertyCaching returns. Therefore, the unit
tests wait for a fixed time period (half a second) after the call to
PBO::EnablePropertyCaching before proceeding with the actual test.

Change-Id: I3ec7c4423ccdbfa8e7544b4aaba6b67703af5888
Signed-off-by: Meggan Ehret <meggan@qeo.org>