ASACORE-1524: Default AuthListener for ECDHE auth mechanisms.
Security 2.0 application requires the app developer to call EnablePeerSecurity
with an AuthListener.  There is no default auth listener, so everybody has to
write one.

This effort is to create a simple default ECDHE AuthListener for developer.
It is very simple by just returning true to the RequestCredentials callback
for ECDHE_NULL and ECDHE_ECDSA key exchanges.  The Security 2.0 framework will
supply the Identity certificate chain to the peer during the ECDHE_ECDSA key
exchange.

The default AuthListener also supports ECDHE_PSK if the app developer
supplies the PSK.

Change-Id: I30c9a994ba61e13ab7149dc6981342edbb0e4815
Signed-off-by: Phil Nguyen <philn@qce.qualcomm.com>