ASACORE-2068 Full conversation hash
Introduce auth version 4 which adds the following to the conversation
hash to protect them:
     * Header bytes for each phase of the conversation
       (ExchangeGuids, ExchangeSuites, etc)
     * Proposed auth protocol versions from each side
     * Session key generation parameters
     * Auth suite exchange

Move conversation hash out of KeyExchanger into PeerState and
introduce the overloaded UpdateHash method, which takes a
conversation version to which a hash applies, and updates the
hash with it if the conversation is happening at at least that
authentication version.

Conversation hashes are now independent between each version,
to improve what is hashed in parts of the conversation currently
hashed in V1, moving away from transformed inputs currently used and
closer to the bytes on the wire.

The hash update functions can fail (they return a QStatus), although
it's unlikely, but there's currently no checking of return codes in
the code. Replumbing all the code to fail gracefully in this case
is too large a change for this work item. Instead, asserts to try to
catch any issues in debug builds.

SampleServiceECDHE is changed to accept all current auth methods for
testing purposes.

Change-Id: Ibfd59953e0385cee1f07d90bc2f7807b979d8301
Signed-off-by: Kevin Kane <kkane@microsoft.com>