ASACORE-2138 ajctest BusAttachmentTest::connect_null can fail on Windows 10
BusAttachmentTest::connect_null compares effective connect spec (when the default
connect spec is used) against the test's connect spec. The effective connect spec
on Windows 10 is TCP when the daemon is running and test is built in "win7" mode
(the current Scons config), thus this test fails in such configuration as it is
different than the test's connect spec, which is to prefer null transport for
win7 mode.

The fix is to change this test to not compare the effective connect spec (after
the call to alljoyn_busattachment_connect) against the test's preferred connect
spec.

I also corrected the macros in BusAttachment.cc and ajtestcommon.cc.

Change-Id: I9d391d08fc85d6dc61a697abba9b810a30822ea3
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>