ASACORE-1872 LanguageTag APIs hard to use since there's no prefix matching
Implement the algorithm defined in RFC 4647 section 3.4.
Asking for "DE-CH" should match "de" or "de-CH" but not "de-CH-1901".
Asking for "fr" when none is found should use the default language.
This reduces extra round trips needed to get localized descriptions.
This change updates GetAboutData() but not IntrospectWithDescription()
which has a different set of problems (it forces the lookup mechanism to be
in the app rather than the framework, so is harder to implement).
Separate JIRA tickets will be used for the other pieces.

Change-Id: I4948c5b9c41317836d9adedca00e5b8b2ad86598
Signed-off-by: Dave Thaler <dthaler@microsoft.com>