ASACORE-2068 Full conversation hash with raw message contents
Instead of recursively parsing messages and hashing the
security-critical fields alone, take the raw data received from or
sent to the network and hash that instead by adding in some internal
plumbing to get the raw content of messages to the authentication
layer code. This still protects the security-critical fields but also
the entire message header as well to protect against possible future
attacks on it. This also greatly simplifies the implementation of
hashing the message contents.

Change-Id: I92857f36f918ee910b4a39eb94afaad65417b3c4
Signed-off-by: Kevin Kane <kkane@microsoft.com>