ASACORE-2182 Do not reset conversation hash after GetSessionKey fail
If the master secret is expired during a key exchange, GetSessionKey
on the responder side will reset the conversation hash, but this
case is not actually fatal, as the initiator will retry. So, we need
to keep the conversation hash alive in this case.

Also add a regression test for this case.

Change-Id: I3d75d947515d9a3e3bb245a93662ed71f62725b2
Signed-off-by: Kevin Kane <kkane@microsoft.com>