ASACORE-804: Bundled Router only configurable in debug mode
Add TEST_CONFIG parameter that sets the pathname of the config file that
the Bundled Router should attempt to load. If not set or if the pathname
is not found at runtime then the Bundled Router will use the default
configuration. This behavior is the same for debug and release VARIANTS.

Example usage:
  scons BR=on TEST_CONFIG="./config.xml"

Change-Id: I4da79ad0c9f5dc32119688c518ae92d8d4d729ae
Signed-off-by: Aaron Pulliam <aaronp@qce.qualcomm.com>