ASACORE-1933: Debug and Release builds should include symbols ASACORE-2173: Remove Windows XP support from AllJoyn build system
The AllJoyn SDK should always include debug symbols for both release and debug
builds to ensure it is easy to investigate issues that customers hit while
using the SDKs. This change will bundle the symbols with the libs for release
builds in the same way that they have been included for debug releases, and it
will include PDB files for all sample EXEs and DLLs in the SDK.

The second part of the change is removing unsupported platforms from the
AllJoyn build system. Windows XP is no longer supported, nor is Visual Studio
9.0 or 10.0.

Change-Id: Ib4643f5dc7e759035e8c8de41af02d2f18af7a3e
Signed-off-by: Rob Smith <robs@microsoft.com>