ASACORE-2231 Use EKUs to indicate certificate type
Update certificate code in various places to use EKUs to indicate if
a certificate is an identity, membership, or unrestricted
certificate instead of a custom extension. Change certificate
ingestion and validation methods to insist leaf certificates are
not unrestricted. Update sample code to incorporate this change.

Correct method that validates a cert chain is valid for a particular
purpose. Correct unit tests for IUI and MUM cases; these are in fact
valid.

Update test cert generation scripts for OpenSSL, and add scripts and
auxiliary files to generate scripts using CAPI on Windows (and then
OpenSSL to post-process the output from the Windows utilities) for
tests in AuthListenerECHDETest.

Add a script to generate certificates and keys with OpenSSL for
SecurityX509Test.

Change-Id: I7127a058eb5aba2a708242ca3a0373645fab0581
Signed-off-by: Kevin Kane <kkane@microsoft.com>