ASACORE-1782 Investigate and remove MD5 Hash algorithm
MD5 is a weak hash algorithm, and is no longer suitable for
cryptographic applications. This change removes MD5 functionality,
but does not remove any public APIs, and thus, it is not deprecation.
This change brings the CNG and OpenSSL implementations in parity
with the built-in crypto implementation, which already returns
an error when MD5 is specified. The change removes dead MD5 code.

There are no known uses of the MD5 algorithm in AllJoyn.

Change-Id: I0fd0db524e13c39616e5c92aed66be12c56c257a
Signed-off-by: Tolga Acar <tolga@microsoft.com>