ASACORE-2228: Hash mismatched when peer auth version is smaller than 4
Fix the conversion hash issue when communication with a peer with auth version
smaller than 4.

Fix the conversion hash of the reply message when the initiator sends back another message before the responder hashes its own reply message.

Disable the signal test in SecurityPolicyRulesTest.cc because of the status
code check of the BusObject::Signal call in the BR=off does not reflect
the permission check since it returns the message submission to the daemon
router.  Will reanble or rewrite these test cases in future commits.

Change-Id: I253519922c4e55c1bf2c2b5280d0736c4484217e
Signed-off-by: Phil Nguyen <philn@qce.qualcomm.com>