ASACORE-2101: Clear out master secrets when identity changed (fixed)
Fixing the timing issue when the master secrets are cleared but the reply
message was not actually delivered yet (even though the BusObject.MethodReply
already returns).  In such case, the encryption fails and the caller is still
waiting for the response.

The fix is to have the sender wait until its reply message is encrypted before
removing the master secrets.

Change-Id: I47f4b62fd95c78efdf9673aed81e7dd9d4e8fa7b
Signed-off-by: Phil Nguyen <philn@qce.qualcomm.com>