ASACORE-2231 Improve generation of CAPI test certificates
Instead of generating test end entity certificates with a basic
constraints extension that does imply the ability to be a CA,
generate these certificates instead with no basic constraints
extension at all; in X509v3, this also is interpreted as an end
entity certificate.

Change the pfx2pem.{sh|cmd} script to use the same filename for
private key files as test_cert_generation.sh, so there's less chance
of the outputs from a run of one comingling with the outputs from a
run of the other. They already use the same filenames for the
certificates.

Update AuthListenerECDHETest's CAPI test with freshly generated
certificates with the updated script.

Change-Id: I2ad5f66aba42de5a1ec44c8fb7a71f163fc54f34
Signed-off-by: Kevin Kane <kkane@microsoft.com>