ASACORE-2183 Memory not securely deleted.
Sensitive data, such as shared secrets, random number states,
secret keys, and private keys must be zeroed before
free/delete/stack-pop to minimize potential exposure in memory
via unintended stack and freed heap access.

The modified code clears memory via calls to ClearMemory() for
DRBG state in the destructor, temporary seed value on the stack,
temporary internal DRBG state on the stack, ECC private key and
computed shared secret in the destructor and on the stack, and
generic key value (KeyBlob) when the Erase() method is called. It
replaces memset() calls that may be optimized out and adds a
couple of constructors with ClearMemory() to clear sensitive
memory to the ECCSecret::ECCSecretState and Crypto_DRBG::Context
structures.

Change-Id: I2d7c8d157fb9aa4bb0b355809d3be8d33893be3e
Signed-off-by: Tolga Acar <tolga@microsoft.com>