ASACORE-2265: Permission check in the signal call before sending to router.
Perform the permission authorization in the BusObject::Signal before sending
the message to the router.  This will allow it to return PERMISSION_DENIED
when the sender is not authorized to emit the signal.  In the BR=off mode,
the permission authorization is not done until the message is to be
encrypted when the router delivers the message.

Change-Id: I37552d653a7be8af23cd903f806d6191013b6126
Signed-off-by: Phil Nguyen <philn@qce.qualcomm.com>