ASACORE-1393 Update the PermissionPolicy ToString method
Updated the ToString Method for the PermissionPolicy and the
KeyInfo classes to output an xml formated string.  Almost all
ToString methods in AllJoyn return xml formated strings. I
found the original output from the ToString methods hard to
read making it difficult understand the contents of the policy
when debugging.

Change-Id: I70b09b77458871591075e61f971615b359258a82
Signed-off-by: George Nash <georgen@qce.qualcomm.com>