ASACORE-2157 Unify algorithm to get localized text
Add translator support to AboutData.
Move the recently-added-but-not-yet-released GetBestLanguage code from
the AboutData class to the Translator class.
Make IntrospectWithDescription use GetBestLanguage to resolve the usability
problem.
Abstract AboutData and IntrospectWithDescription translation code into
generic StringTableTranslator and MsgArgTranslator classes to facilitate
reuse.
Make IntrospectWithDescription support a default translator like AboutData
already did, using StringTableTranslator.

Note that C binding support is not part of this ticket, since it wasn't
supported for translation before this change.  That support is tracked by
ASACORE-2174.  Also note that the whole reason for the MsgArgTranslator
comes from the poorly designed AboutData::GetField() API which requires
AboutData to internally keep a table of MsgArg structures it can return.
In the future, this could possibly be deprecated (tracked by
ASACORE-2323), but in this change it must be supported so that no existing
code will break that may rely on it.

Change-Id: I59a36c037e4c2fcede5c7646131a3bd485387390
Signed-off-by: Dave Thaler <dthaler@microsoft.com>