ASACORE 2287: Use certificates with EKUs in bbclient and bbservice
Now that certificate type is encoded as an EKU, the test certificates
in bbclient and bbservice must also have EKUs.  This is related to
ASACORE-2287, but I'm not sure it's the right resolution, I have to
follow-up with the reporter. This fixes the problem caused by the
repro steps, and should be done regardless.

Change-Id: I3bd3a9e63a86decaceba02517c661ad0a5db414d
Signed-off-by: Greg Zaverucha <gregz@microsoft.com>