ASACORE 2287: bbclient and bbservice certificates are rejected
Now that certificate type is encoded as an EKU, and KeyExchanger was
changed to enforce presence of an EKU on the EE cert, older
certificates are rejected.  This change relaxes EKU checking in
ECDHE_ECDSA to allow EE certs with no EKU.  Older certificates with
a non AllJoyn EKU will still be rejected, however.

Change-Id: I3bd3a9e63a86decaceba02517c661ad0a5db414d
Signed-off-by: Greg Zaverucha <gregz@microsoft.com>