ASACORE-2388: App can hang during install identity
When one thread clears the peer secrets from the keystore and
another thread tries to store the application state in the same
keystore, there is a hang because of deadlock on the Mutex and
the keystore's sink lock.

Change the KeyStore::Clear method to use an private method DeleteKey
which does not acquire additional Mutex lock.

Modify KeyStore::DelKey to use DeleteKey as well.

Change-Id: I93f5d180d9316f559b4f5d5a1f7848c721a3331f
Signed-off-by: Phil Nguyen <philn@qce.qualcomm.com>