ASACORE-2409: AllJoynPeerObj lock being held too long
The AllJoynPeerObj global lock is held too long in order to the protect
the order of the key exchange message hashing process.

A different mechanism to ensure proper order of the key exchange message
hashing mechanism is created.

Change the ObserverTest to use in memory keystore so fix the build error.

Change-Id: I2decd14feb393e0f7d1d5a620ddb7ed0bfca45bf
Signed-off-by: Phil Nguyen <philn@qce.qualcomm.com>