ASACORE-2282 Documentation for UpdatePolicy and UpdateIdentity
This adds documentation to the UpdatePolicy and UpdateIdentity member function.  It tells the user that calling these methods
will cause the target peer to clear its keys.  It also clarifies
that ProxyBusObject.SecureConnection(true) can be called to
force the peers to create a new set of keys.

Also updated the unit tests to follow the recommendation given
in the API documentation.  They no longer make a call to
GetPolicy to force reauthentication.

Change-Id: If75d51abc35ee8ac7fabab33ca410ef0c35461ab
Signed-off-by: George Nash <georgen@qce.qualcomm.com>