ASACORE-2406 consistent behavior for sessioncast signals on point-to-point sessions
If security is disabled, a sessioncast signal on a point-to-point
session is functionally equivalent to a unicast signal over the same
session. When security (in particular Security 2.0) is enabled, this is
no longer the case: for a unicast signal, policies are enforced, for a
sessioncast signal (which may have multiple recipients in a multipoint
session) policies are not enforced (or not in the same way).

This patch aims to restore consistency for sessioncast signals on
point-to-point sessions. To achieve this, the existing session
bookkeeping in the BusAttachment is extended to keep track of all
participants in a session. If it's a p2p session, sessioncast signals
are transparently rewritten as unicast signals before emission.

Change-Id: I236cac907bb916daac9c8dafedcd9f0997114622
Signed-off-by: Meggan Ehret <meggan@qeo.org>