ASACORE-2038: add "reference" TOC (SDK "manifest") files to source tree
ASACORE-2038 "SDK Configuration: Define a manifest and add a check to
identify delta" is implemented in two parts:
1. Adding some initial "reference" manifests (now dubbed "TOC" files)
into the AllJoyn Core source tree core/alljoyn.git
2. Adding some scripting to every Jenkins job that builds an SDK, to
compare the contents of the newly-built SDK to that of the "reference".

This commit is part 1. This commit simply adds the new reference files,
and will not affect any existing function or feature in any way.

Part 2 is implemented as changes to Ant and other build scripts, in
extras/automation.git.

Change-Id: I9624466f6e2d57d44862b089baa9cdc96167d563
Signed-off-by: Tony Wallace <awallace@qce.qualcomm.com>