ASACORE-2528 NamedPipeDaemonTransport::Join can get stuck preventing shutdown
Windows AllJoyn service can get stuck during shutdown because not all
the NamedPipeDaemonEndpoints are accounted for, per this callstack:

    MSAJApi!ajn::DaemonTransport::Join
    MSAJApi!ajn::NamedPipeDaemonTransport::Join
    MSAJApi!ajn::TransportList::Join
    MSAJApi!ajn::BusAttachment::Internal::TransportsJoin
    MSAJApi!ajn::BusAttachment::WaitStopInternal
    MSAJApi!ajn::BusAttachment::Join
    MSAJApi!ajn::BusController::Join
    MSAJApi!ajn::BusController::~BusController

The issue is that the NamedPipeDaemonTransport doesn't properly get
notified when an *authenticating* endpoint goes away. The result is that its
internal list (NamedPipeDaemonTransport::authenticatingEndpointList)
can contain an invalidated endpoint, resulting in Join getting stuck in the
clean up path.

The fix is to introduce a callback in NamedPipeDaemonTransport to clean
up *authenticating* endpoint list as well as to ensure that the endpoints
that have been invalidated do not make their way into
DaemonTransport::endpointList.

Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>
Change-Id: Ife06aad80f0d0ecf9582fb4aad8275af4d32949d