ASACORE-2537 Fix NamedPipeDaemonEndpoints leaks
NamedPipeDaemonEndpoint objects are leaking memory when app goes away during
authentication or when the daemon is unloaded. As a NamedPipeDaemonEndpoint
goes away, the entry is removed from DaemonTransport object, but without
Join() being called. NamedPipeDaemonEndpoint requires that Join() is called
in order to properly clean itself up. This is due to the fact that a lifetime
reference count is taken before the authentication and is released by Join().

Change-Id: I275528506c24cf8cca663e681200b73f31e0f8a1
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>