ASACORE-2537 Fix NamedPipeDaemonEndpoints leaks
NamedPipeDaemonEndpoint objects are leaking memory when apps go away.
As an NamedPipeDaemonEndpoint goes away, the entry is removed from the
list on DaemonTransport object, but without Join() being called.
NamedPipeDaemonEndpoint relies on Join() to be called in order to properly
clean itself up. This is due to the fact that the ref count is taken
before authentication and is expected to be decremented by Join().

Change-Id: I275528506c24cf8cca663e681200b73f31e0f8a1
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>