ASACORE-2497: SecSignalTest.SecureConnection intermittently failed
The SecSignalTest.SecureConnection intermittently failed because of the
asynchronous reload of the shared keystore within the same process by the
concurrent key exchange threads in the AllJoynPeerObj.

The keystore code is modified to support an consistency lock.
The KeyStore::Store call will acquire the consistency lock, reload, and then
call the Listener::StoreRequest.  This would allow multiple threads in the same
process to reload the shared keystore.

The keys load is now loaded to a different hash table in order to prevent
wiping out new entries added to the keys hash table by other threads while
the load is in progress.

Change-Id: I485a03b838e145fbbfac3db5a2e5c0675a44616d
Signed-off-by: Phil Nguyen <philn@qce.qualcomm.com>