ASACORE-2386 Add a platform independent QCC_ASSERT macro (part 1)
Add a QCC_ASSERT macro to platform.h (C++) and AjAPI.h (C),
and remove nearly all references to <assert.h> and the assert macro
in C and C++ code. This will allow future redefinition of the macro
to platform-specific alternatives, to allow customizing the behavior
of failed assertions from the assert macro's default, such as
dropping into a debugger, or emitting telemetry.

JavaScript binding is unaffected by this change.

A couple of instances remain where unit tests depend on the precise
behavior of the assert macro. DiscoveryDeathTest depends on the
use of assert in alljoyn_core\router\ns\IpNameServiceImpl.cc.

Two incorrect uses of assert are fixed in
alljoyn_core\unit_test\ClientSetup.cc.

A pragma in qcc\platform.h that only applies in C++ mode is amended
to avoid a warning when building in pure C mode.

Change-Id: Ia06ed92009554e2410fc4c32f1778e497ed49225
Signed-off-by: Kevin Kane <kkane@microsoft.com>