ASACORE-2386 Default QCC_ASSERT to _ASSERT on Windows
On Windows, prefer the _ASSERT macro as the implementation of
QCC_ASSERT. A pre-existing definition for QCC_ASSERT will continue
to prevail if present.

The DiscoveryDeathTest unit test suite depends on the particular
behavior of the assert() macro which is no longer appropriate, now
that QCC_ASSERT is changeable. Since the APIs being tested are
internal, change them to instead check the asserted conditions and
return an error code if they don't hold, and instead test for that
failed error code instead of an assertion failure and process abort.

Change-Id: I22908fecd5c94523d09397bd129f0a75c52a7e3f
Signed-off-by: Kevin Kane <kkane@microsoft.com>