ASACORE-2081 Security 2.0 Fixed Security other unit test
Some of the Test were sending a signal from a Security1.0 peer
to a Security1.0 peer when they should have been sending from a
Security2.0 peer.  Error was missed in code review and caught
when porting the tests to thin core library.

Since the Security2.0 peer can not get the manifest from the
security1.0 peer the permission denied error makes since and
matches the behavior seen in the TC tests.

Change-Id: Idd7431a08b1cafe431172c863b0ea9c718eebca5
Signed-off-by: George Nash <georgen@qce.qualcomm.com>