Updates to the AllJoyn documentation.
This moves the documentation that was found in the
allJoynDocumentation file into two files.
  AllJoynDocumentation.md
  alljoyn_namespace.dox

The document is now in markdown format that is simpler
to read and modify. The namespace delcarations that
still needed to be in a code style format were moved to
alljoyn_namespace.dox file.

This also add the messages_and_msgarg documentation. This
is a collection of documentation with corresponding code
snippets that show how build build almost any container type
a developer would want to create using the MsgArg class.

The code snippet code is build by the compiler if the developer
is producing documentation.  This is done so any code changes
that affect the documentation can be discovered since the
compile will fail.

Change-Id: Ib84b1ca283d5fa4bc2e2fd701baac436f9456d0b
Signed-off-by: George Nash <georgen@qce.qualcomm.com>